fix(ast/windows): 7 correctness issues found in review  High: - ScriptBlockExpressionAst no longer emits ReasonSubshell.  { } blocks are   idiomatic PowerShell parameter syntax (Where-Object, ForEach-Object, etc.)   and must not be treated as subshell execution.  Invoke-Command is already   caught by invokeRiskCmdlets; the standalone script-block node adds nothing.  Medium: - policy_test.go: add ReasonDestructive to soft-signal loop coverage   (it was added to the loop but not tested). - roundTrip timeout comment: clarified that closing stdin does not immediately   unblock ReadBytes; the goroutine unblocks when the bridge process terminates.  Low: - startBridgeProcess: close stdinPipe on StdoutPipe failure or Start() failure   to avoid resource leak. - snapshot_test.go: split Windows corpus/helper/test into   snapshot_windows_test.go with //go:build windows tag so the test actually   runs (not silently skipped) on Linux/Mac CI. - TestWindowsParser_BridgeRestart: call cmd.Wait() in a goroutine after   Process.Kill() for correct process cleanup. - ps_bridge.ps1: document sc alias collision with sc.exe (Service Control   Manager) inline in the destructiveCmdlets list.  Tests added: - BridgeContract: Where-Object { } and ForEach-Object { } cases verify no   ReasonSubshell flag. - WindowsCorpus: ForEach-Object { } entry confirms policy-level auto-approve   for operator-only commands between allowlisted cmdlets.
